requirements.txt
Babel==2.9.1
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
SQLAlchemy==1.4.22
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
